Dockerfile: Split into two-stage build
Building on i686 revealed that certain dependencies must be compiled
during installation on uncommon platforms; this wasn't noticed on x86_64
or aarch64, but was needed on i686, and I expect this to be the case on
other platforms that aren't "normal". Splitting the docker build process
into two stages allows me to handle that without doing cleanup and
possibly uploading unnecessary layers.
To reduce the possibility of runtime issues, the gunicorn-run target no
longer depends on the requirements target, as I don't know exactly how
pipenv behaves when a dependency has a newer version available than
what's specified in Pipfile.lock