RPC.util.base.Api: More robust caller detection
The code is ugly as all fuck and i'm pretty sure the former python bdfl
Guillermo Del Toro or whoever the fuck his name is, is organising a
sniper hit squad to kill me for writing this code, but whatever.
Flask's blueprint stuff is neat but my subclass of it that autodetects
the information from the stack is way better.